First_laba_try_10

+++ tasks/moskaev_v_max_value_elem_matrix/mpi/src/ops_mpi.cpp
@@ -42,7 +42,7 @@ bool MoskaevVMaxValueElemMatrixMPI::RunImpl() {
   auto remainder = total_rows % size;
 
   int start_row = (rank * rows_per_process) + std::min(rank, static_cast<int>(remainder));
-  int end_row = start_row + rows_per_process + (rank < remainder ? 1 : 0);
+  int end_row = start_row + rows_per_process + (rank < static_cast<int>(remainder) ? 1 : 0);
 
   // Поиск локального максимума в своей части матрицы
   int local_max = matrix[start_row][0];
